Fix confidence interval calculations for ROCs
Calculating confidence interval regions for ROC curves is highly
non-trivial: one possibility is calculating the TPR-confidence intervals
for fixed FPR points.  However, this has some unintuitive consequences:

Consider the simple case of TPR/FPR for a binary classification task:
The AUROC should be the same, independently of which class we use as the
positive label.  However, since the TPR/FPR axes get switched, confidence
intervals calculated this may differ.

Additionally, the AUROC under the upper / lower confidence ROC curves
calculated this way is different from the confidence intervals of the
AUROCs themselves, as a displacement of the ROC curve has a quadratic
effect on the AUROC.

Finally, calculating upper / lower ROC curves this way may lead to
curves which have points implying a TPR outside of the [0, 1] interval.

Thus, the graphical representation of the confidence intervals has been
removed.